+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [Unreleased]
+
+### Added
+
+- Add option --dry-run, add option --no-enforce-prefix
+
 ## [2.0.1] - 2024-05-03
 
 ### Added
@@ -53,6 +59,7 @@
 .github/workflows to make github installing flutter and not dart SDK'
 - Prepare publish
 
+[Unreleased]: https://github.com/inlavigo/gg_create_package/compare/2.0.1...HEAD
 [2.0.1]: https://github.com/inlavigo/gg_create_package/compare/2.0.0...2.0.1
 [2.0.0]: https://github.com/inlavigo/gg_create_package/compare/1.0.3...2.0.0
 [1.0.3]: https://github.com/inlavigo/gg_create_package/compare/1.0.2...1.0.3